fake histogram in race mode as hdrhistogram has apparent races and is not under active development
