+++ instrumentation/metric/histogram.go
@@ -1,3 +1,4 @@
+//+build !race
 // Copyright 2019 the orbs-network-go authors
 // This file is part of the orbs-network-go library in the Orbs project.
 //
@@ -9,8 +10,6 @@ package metric
 import (
 	"fmt"
 	"github.com/codahale/hdrhistogram"
-	"github.com/orbs-network/scribe/log"
-	"strconv"
 	"sync/atomic"
 	"time"
 )
@@ -19,25 +18,6 @@ type Histogram struct {
 	namedMetric
 	histo         *hdrhistogram.WindowedHistogram
 	overflowCount int64
-}
-
-type histogramExport struct {
-	Name    string
-	Min     float64
-	P50     float64
-	P95     float64
-	P99     float64
-	Max     float64
-	Avg     float64
-	Samples int64
-}
-
-func toMillis(nanoseconds int64) float64 {
-	return floatToMillis(float64(nanoseconds))
-}
-
-func floatToMillis(nanoseconds float64) float64 {
-	return nanoseconds / 1e+6
 }
 
 func newHistogram(name string, max int64, n int) *Histogram {
@@ -100,41 +80,4 @@ func (h *Histogram) Export() exportedMetric {
 
 func (h *Histogram) Rotate() {
 	h.histo.Rotate()
-}
-
-func (h histogramExport) LogRow() []*log.Field {
-	if h.Samples == 0 {
-		return nil
-	}
-
-	return []*log.Field{
-		log.String("metric", h.Name),
-		log.String("metric-type", "histogram"),
-		log.Float64("min", h.Min),
-		log.Float64("p50", h.P50),
-		log.Float64("p95", h.P95),
-		log.Float64("p99", h.P99),
-		log.Float64("max", h.Max),
-		log.Float64("avg", h.Avg),
-		log.Int64("samples", h.Samples),
-	}
-}
-
-func (h histogramExport) PrometheusRow() []*prometheusRow {
-	name := h.PrometheusName()
-	return []*prometheusRow{
-		{name, 0.01, strconv.FormatFloat(h.Min, 'f', -1, 64)},
-		{name, 0.5, strconv.FormatFloat(h.Min, 'f', -1, 64)},
-		{name, 0.95, strconv.FormatFloat(h.Min, 'f', -1, 64)},
-		{name, 0.99, strconv.FormatFloat(h.Min, 'f', -1, 64)},
-		{name, 0.999, strconv.FormatFloat(h.Min, 'f', -1, 64)},
-	}
-}
-
-func (h histogramExport) PrometheusType() string {
-	return "histogram"
-}
-
-func (h histogramExport) PrometheusName() string {
-	return prometheusName(h.Name)
 }
